+++ src/components/common/Feedback/feedbackTypes.ts
@@ -1,0 +1,5 @@
+export interface IAverages {
+  score1: number;
+  score2: number;
+  score3: number;
+}
