+++ src/styles/components/common/fullscreenImage.scss/imageInfo.scss
@@ -48,8 +48,8 @@
     .user {
       font-weight: 200;
       font-size: 2.4rem;
-      margin-bottom: 0.5rem;
-      font-family: $display-font;
+      margin-bottom: 2rem;
+      font-family: $text-font;
       letter-spacing: 0.05rem;
       line-height: 2rem;
     }
@@ -57,8 +57,7 @@
       font-weight: 200;
       font-size: 1.6rem;
       strong {
-        font-weight: 500;
-        color: $red;
+        font-weight: 700;
       }
     }
   }
