+++ src/styles/components/common/submission.scss
@@ -1,3 +1,4 @@
+// TODO split the styles for feedback and submission
 .submission-details {
   @include wrapped-container();
   background-color: $w-grey-10;
